+++ src/Symfony/Component/Serializer/Normalizer/UidNormalizer.php
@@ -13,6 +13,10 @@ namespace Symfony\Component\Serializer\Normalizer;
 
 use Symfony\Component\Serializer\Exception\LogicException;
 use Symfony\Component\Serializer\Exception\NotNormalizableValueException;
+use Symfony\Component\Serializer\InvariantViolation;
+use Symfony\Component\Serializer\Result\DenormalizationResult;
+use Symfony\Component\Serializer\Result\NormalizationResult;
+use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Uid\AbstractUid;
 use Symfony\Component\Uid\Ulid;
 use Symfony\Component\Uid\Uuid;
@@ -42,6 +46,17 @@ final class UidNormalizer implements NormalizerInterface, DenormalizerInterface,
      */
     public function normalize($object, string $format = null, array $context = [])
     {
+        $result = $this->doNormalize($object, $context);
+
+        if ($context[SerializerInterface::RETURN_RESULT] ?? false) {
+            return NormalizationResult::success($result);
+        }
+
+        return $result;
+    }
+
+    public function doNormalize(AbstractUid $object, array $context): string
+    {
         switch ($context[self::NORMALIZATION_FORMAT_KEY] ?? $this->defaultContext[self::NORMALIZATION_FORMAT_KEY]) {
             case self::NORMALIZATION_FORMAT_CANONICAL:
                 return (string) $object;
@@ -70,10 +85,24 @@ final class UidNormalizer implements NormalizerInterface, DenormalizerInterface,
     public function denormalize($data, string $type, string $format = null, array $context = [])
     {
         try {
-            return Ulid::class === $type ? Ulid::fromString($data) : Uuid::fromString($data);
+            $result = Ulid::class === $type ? Ulid::fromString($data) : Uuid::fromString($data);
         } catch (\InvalidArgumentException $exception) {
-            throw new NotNormalizableValueException(sprintf('The data is not a valid "%s" string representation.', $type));
+            $message = sprintf('The data is not a valid "%s" string representation.', $type);
+
+            if ($context[SerializerInterface::RETURN_RESULT] ?? false) {
+                $violation = new InvariantViolation($data, $message, $exception);
+
+                return DenormalizationResult::failure(['' => [$violation]]);
+            }
+
+            throw new NotNormalizableValueException($message);
         }
+
+        if ($context[SerializerInterface::RETURN_RESULT] ?? false) {
+            return DenormalizationResult::success($result);
+        }
+
+        return $result;
     }
 
     /**