+++ src/Symfony/Component/Serializer/Tests/Normalizer/DataUriNormalizerTest.php
@@ -14,6 +14,9 @@ namespace Symfony\Component\Serializer\Tests\Normalizer;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\Serializer\Normalizer\DataUriNormalizer;
+use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
+use Symfony\Component\Serializer\Result\DenormalizationResult;
+use Symfony\Component\Serializer\SerializerInterface;
 
 /**
  * @author Kévin Dunglas <dunglas@gmail.com>
@@ -167,6 +170,34 @@ class DataUriNormalizerTest extends TestCase
             ['data:a!b#c&d-e^f_g+h.i/a!b#c&d-e^f_g+h.i;base64,foobar'],
             ['data:text/plain;charset=utf-8;base64,SGVsbG8gV29ybGQh'],
         ];
+    }
+
+    /**
+     * @dataProvider validUriProvider
+     */
+    public function testItDenormalizesAndReturnsSuccessResult($uri): void
+    {
+        $result = $this->normalizer->denormalize($uri, \SplFileObject::class, null, [
+            SerializerInterface::RETURN_RESULT => true,
+        ]);
+
+        self::assertInstanceOf(DenormalizationResult::class, $result);
+        self::assertTrue($result->isSucessful());
+        self::assertInstanceOf(\SplFileObject::class, $result->getDenormalizedValue());
+    }
+
+    public function testItDenormalizesAndReturnsFailureResult(): void
+    {
+        $result = $this->normalizer->denormalize('not-a-uri', \SplFileObject::class, null, [
+            SerializerInterface::RETURN_RESULT => true,
+        ]);
+
+        self::assertInstanceOf(DenormalizationResult::class, $result);
+        self::assertFalse($result->isSucessful());
+        self::assertSame(
+            ['' => ['The provided "data:" URI is not valid.']],
+            $result->getInvariantViolationMessages()
+        );
     }
 
     private function getContent(\SplFileObject $file)