+++ src/Symfony/Component/Serializer/Tests/Normalizer/UidNormalizerTest.php
@@ -4,7 +4,10 @@ namespace Symfony\Component\Serializer\Tests\Normalizer;
 
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\Serializer\Exception\LogicException;
+use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
 use Symfony\Component\Serializer\Normalizer\UidNormalizer;
+use Symfony\Component\Serializer\Result\DenormalizationResult;
+use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Uid\AbstractUid;
 use Symfony\Component\Uid\Ulid;
 use Symfony\Component\Uid\Uuid;
@@ -167,5 +170,52 @@ class UidNormalizerTest extends TestCase
         $this->normalizer->normalize(new Ulid(), null, [
             'uid_normalization_format' => 'ccc',
         ]);
+    }
+
+    /**
+     * @dataProvider dataProvider
+     */
+    public function testItDenormalizesAndReturnsSuccessResult(string $uuidString, string $class): void
+    {
+        $result = $this->normalizer->denormalize($uuidString, $class, null, [
+            SerializerInterface::RETURN_RESULT => true,
+        ]);
+
+        self::assertInstanceOf(DenormalizationResult::class, $result);
+        self::assertTrue($result->isSucessful());
+        self::assertEquals(
+            Ulid::class === $class ? new Ulid($uuidString) : Uuid::fromString($uuidString),
+            $result->getDenormalizedValue()
+        );
+    }
+
+    /**
+     * @dataProvider failureDataProvider
+     */
+    public function testItDenormalizesAndReturnsFailureResult(string $class): void
+    {
+        $result = $this->normalizer->denormalize('not-an-uuid', $class, null, [
+            SerializerInterface::RETURN_RESULT => true,
+        ]);
+
+        self::assertInstanceOf(DenormalizationResult::class, $result);
+        self::assertFalse($result->isSucessful());
+        self::assertSame(
+            ['' => [sprintf('The data is not a valid "%s" string representation.', $class)]],
+            $result->getInvariantViolationMessages()
+        );
+    }
+
+    public function failureDataProvider(): iterable
+    {
+        return [
+            [UuidV1::class],
+            [UuidV3::class],
+            [UuidV4::class],
+            [UuidV5::class],
+            [UuidV6::class],
+            [AbstractUid::class],
+            [Ulid::class],
+        ];
     }
 }