+++ src/Symfony/Component/Serializer/CHANGELOG.md
@@ -15,6 +15,7 @@ CHANGELOG
  * added `FormErrorNormalizer`
  * added `MimeMessageNormalizer`
  * serializer mapping can be configured using php attributes
+ * added `SerializerInterface::RETURN_RESULT` context option to collect (de)normalization errors instead of throwing immediately
 
 5.1.0
 -----